#VP-6436 add BaseOutlinePartResolver
