+++ src/VirtoCommerce.CatalogModule.Core/Model/OutlinePart/BaseOutlinePartResolver.cs
@@ -1,0 +1,24 @@
+using VirtoCommerce.Platform.Core.Common;
+
+namespace VirtoCommerce.CatalogModule.Core.Model.OutlinePart
+{
+    public abstract class BaseOutlinePartResolver : IOutlinePartResolver
+    {
+        public abstract string ResolveOutlinePart(IEntity entity);
+
+        public string ResolveOutlineName(IEntity entity)
+        {
+            var result = entity.Id;
+            if (entity is Catalog catalog)
+            {
+                result = catalog.Name;
+            }
+            else if (entity is Category category)
+            {
+                result = category.Name;
+            }
+
+            return result;
+        }
+    }
+}
