+++ src/VirtoCommerce.CatalogModule.Core/Model/OutlinePart/CodeOutlinePartResolver.cs
@@ -5,9 +5,9 @@ namespace VirtoCommerce.CatalogModule.Core.Model.OutlinePart
     /// <summary>
     /// Uses codes for outline items.
     /// </summary>
-    public class CodeOutlinePartResolver : IOutlinePartResolver
+    public class CodeOutlinePartResolver : BaseOutlinePartResolver, IOutlinePartResolver
     {
-        public string ResolveOutlinePart(IEntity entity)
+        public override string ResolveOutlinePart(IEntity entity)
         {
             // Fall-back to id.
             var result = entity.Id;
@@ -22,19 +22,6 @@ namespace VirtoCommerce.CatalogModule.Core.Model.OutlinePart
             return result;
         }
 
-        public string ResolveOutlineName(IEntity entity)
-        {
-            var result = entity.Id;
-            if (entity is Catalog catalog)
-            {
-                result = catalog.Name;
-            }
-            else if (entity is Category category)
-            {
-                result = category.Name;
-            }
-
-            return result;
-        }
+        
     }
 }
